Модулирует проект
