+++ js/util.js
@@ -1,0 +1,13 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const checkStringLength = (string, length) => string.length <= length;
+
+const getRandomElement = (array) =>
+  array[getRandomInteger(0, array.length - 1)];
+
+export { getRandomInteger, checkStringLength, getRandomElement};
